pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "nthlayer-correlate"
version = "0.1.0a1"
description = "SitRep — Situational awareness through automated signal correlation"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
dependencies = [
    "pyyaml>=6.0.1",
    "structlog>=24.1",
    "anthropic>=0.39",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.8",
]

[project.scripts]
nthlayer-correlate = "nthlayer_correlate.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
